Add support for *, +, #, -, . and space in codes

+++ app/src/main/java/se/oscarb/pinapple/Crypto.java
@@ -8,7 +8,8 @@ public interface Crypto {
 
     // Encrypt and decrypt with key and message as int
     int encrypt(int messagePlainText, int key);
-    int decrypt(int messageEncrypted, int key);
+
+    long decrypt(long messageEncrypted, int key);
 
 
 }
